This is the latest version of the gsrs Excel tools
it has 2 new scripts: 'Touch' which updates an existing record without
changing any data and 'Replace Name' which replaces the text part of an
existing name entry in an existing substance.

The VBA is under construction to provide a check of ongoing callbacks.
This is not complete yet.
The JavaScript is under construction to provide a clear message when a
resolution operation fails.